pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "AGISwarm.asyncio-queue-manager"
dynamic = ["version"]
description = "asyncio-queue-manager is a Python package that provides a robust and flexible solution for managing asynchronous queues in applications using asyncio. It offers an easy-to-use interface for handling task queues with advanced features such as priority scheduling, concurrency control, and error handling."
readme = "README.md"

requires-python = ">=3.10"
license = { file = "LICENSE.txt" }
keywords = ["sample", "setuptools", "development"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    'ulid-py',
    'sortedcontainers',
]
[project.optional-dependencies]
test = ['pytest', 'pytest-asyncio']
analyze = ['pyright', 'pylint', 'bandit', 'black', 'isort']
build = ['setuptools', 'wheel', 'build']
publish = ['twine']


[tool.setuptools.dynamic]
version = { file = "VERSION" }


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
asyncio_queue_manager = ["py.typed"]
